Add GitHub Actions workflows for testing, releasing, and common steps (#4)
- Added .github/workflows/common.yml:
This file contains reusable workflow steps for setting up the Go environment and running tests. It is referenced by other workflows to maintain consistency and reduce duplication.

- Added .github/workflows/tests.yml:
This file defines a workflow to run tests on all push and pull request events for all branches. It ensures that code changes do not break existing functionality by running tests before merging or deploying.

- Added .github/workflows/release.yml:
This file defines a workflow to create a release when a new tag is pushed. It builds the application for multiple platforms, creates release artifacts, updates the CHANGELOG.md, and publishes Docker images to Docker Hub and GitHub Container Registry.